refactor: 핸들러 함수 모듈화 channelCreate 함수 모듈화
