+++ src/handlers/channel/channelCreatePermissions.js
@@ -1,0 +1,26 @@
+const channelCreatePermissions = async (channel) => {
+  try {
+    // 채널을 만든 사용자의 정보를 가져옵니다.
+    const logs = await channel.guild.fetchAuditLogs({
+      limit: 1,
+      type: 10,
+    });
+
+    const creatorId = logs.entries.first().executor.id;
+    const creator = await channel.guild.members.fetch(creatorId);
+
+    // 생성자에게만 'MANAGE_CHANNELS' 권한 부여
+    await channel.permissionOverwrites.edit(creator.id, {
+      ManageChannels: true,
+    });
+
+    // @everyone에 대한 'MANAGE_CHANNELS' 권한 제거
+    await channel.permissionOverwrites.edit(channel.guild.roles.everyone, {
+      ManageChannels: false,
+    });
+  } catch (error) {
+    console.error('Error setting permissions:', error);
+  }
+};
+
+export default channelCreatePermissions;
