+++ src/index.js
@@ -1,6 +1,7 @@
 import { REST, Routes, Client, GatewayIntentBits } from 'discord.js';
 import 'dotenv/config';
 import readyPermissions from './handlers/channel/readyPermissions.js';
+import channelCreatePermissions from './handlers/channel/channelCreatePermissions.js';
 
 // for Refactor
 const commands = [
@@ -38,28 +39,7 @@ client.on('interactionCreate', async (interaction) => {
 });
 // refactor
 client.on('channelCreate', async (channel) => {
-  try {
-    // 채널을 만든 사용자의 정보를 가져옵니다.
-    const logs = await channel.guild.fetchAuditLogs({
-      limit: 1,
-      type: 10,
-    });
-
-    const creatorId = logs.entries.first().executor.id;
-    const creator = await channel.guild.members.fetch(creatorId);
-
-    // 생성자에게만 'MANAGE_CHANNELS' 권한 부여
-    await channel.permissionOverwrites.edit(creator.id, {
-      ManageChannels: true,
-    });
-
-    // @everyone에 대한 'MANAGE_CHANNELS' 권한 제거
-    await channel.permissionOverwrites.edit(channel.guild.roles.everyone, {
-      ManageChannels: false,
-    });
-  } catch (error) {
-    console.error('Error setting permissions:', error);
-  }
+  await channelCreatePermissions(channel);
 });
 
 // DISCORD 내 BOT TOKEN을 setToken의 인수로 넣어주세요.
